[refactor] refactoring Landing Page file structure
